+++ apps/server/test/hashcash.e2e.test.ts
@@ -38,7 +38,7 @@ describe('hashcash challenge flow', () => {
 
   beforeAll(async () => {
     tmp = mkdtempSync(join(tmpdir(), 'faucet-hc-'));
-    const config = ServerConfigSchema.parse({
+    const config = ServerConfigSchema.parse({ geoipBackend: "none",
       network: 'test',
       dataDir: tmp,
       signerDriver: 'rpc',